Extract OTLP text logging from logging exporter
The plan is to eventually use this as an unmarshaller (#3044) but for now make
it accessible from `internal/otlptext`.
