+++ exporter/loggingexporter/logging_exporter_test.go
@@ -22,7 +22,6 @@ import (
 	"go.uber.org/zap"
 
 	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.opentelemetry.io/collector/internal/testdata"
 )
 
@@ -61,36 +60,4 @@ func TestLoggingLogsExporterNoErrors(t *testing.T) {
 	assert.NoError(t, lle.ConsumeLogs(context.Background(), testdata.GenerateLogDataOneEmptyLogs()))
 
 	assert.NoError(t, lle.Shutdown(context.Background()))
-}
-
-func TestNestedArraySerializesCorrectly(t *testing.T) {
-	ava := pdata.NewAttributeValueArray()
-	ava.ArrayVal().AppendEmpty().SetStringVal("foo")
-	ava.ArrayVal().AppendEmpty().SetIntVal(42)
-
-	ava2 := pdata.NewAttributeValueArray()
-	ava2.ArrayVal().AppendEmpty().SetStringVal("bar")
-	ava2.CopyTo(ava.ArrayVal().AppendEmpty())
-
-	assert.Equal(t, 3, ava.ArrayVal().Len())
-	assert.Equal(t, "[foo, 42, [bar]]", attributeValueToString(ava))
-}
-
-func TestNestedMapSerializesCorrectly(t *testing.T) {
-	ava := pdata.NewAttributeValueMap()
-	av := ava.MapVal()
-	av.Insert("foo", pdata.NewAttributeValueString("test"))
-
-	ava2 := pdata.NewAttributeValueMap()
-	av2 := ava2.MapVal()
-	av2.InsertInt("bar", 13)
-	av.Insert("zoo", ava2)
-
-	expected := `{
-     -> foo: STRING(test)
-     -> zoo: MAP({"bar":13})
-}`
-
-	assert.Equal(t, 2, ava.MapVal().Len())
-	assert.Equal(t, expected, attributeValueToString(ava))
 }
